+++ api/views/__init__.py
@@ -1,2 +1,8 @@
-from .team_related_views import *
 from .social_login_views import *
+from .ticket import *
+from .event import *
+from .tag import *
+from .status import *
+from .member import *
+from .pinned_ticket import *
+from .team import *
